refactor: address smncd review comments on PR #56
- Split FlowRecovery.ts: types stay in types/, utilities move to flowRecoveryUtils.ts
- Rename FlowStateManager → FlowStateStore (backward-compat alias kept)
- Rename RetryTransportWrapper → TransportWithRetry, move to decorators/
- Fix off-by-one: attemptNumber < effectiveMaxRetries → <=
- Remove dead-end prepareFlowRetry() from transport wrapper
- Add cleanupExpired() call in FlowStateStore.create()
- Remove unnecessary re-export of RetryConfig
- Keep FlowErrorCodes as `as const` (better tree-shaking than enum)
- Add backward-compatible type aliases for renamed exports

All 49 tests pass.
